sistemate alcune immagini

+++ domain/src/main/java/com/unibo/android/domain/models/FilmEntity.kt
@@ -12,7 +12,7 @@ data class FilmEntity(
     val genere: String,
     val durata: String,
     val regista: String,
-    val punteggio: Int,
+    val punteggio: Double,
     val percorsoLocandina: String,
     val preferito: Boolean = false
 )
